+++ bin/setup_tests
@@ -1,5 +1,7 @@
 #!/bin/bash
 if [ "$TRAVIS_BRANCH" = "DevelopmentBranch" ]; then
 	sleep $HEROKU_BUILD_TIME
+else
+	rm -rf src/test/java/is/teamgoomba/TicTacToe/Selenium
 fi
 exit 0
