+++ src/test/java/is/teamgoomba/TicTacToe/Selenium/SeleniumTest.java
@@ -1,0 +1,40 @@
+package is.teamgoomba.TicTacToe.Selenium;
+
+import java.util.regex.Pattern;
+import java.util.concurrent.TimeUnit;
+import org.junit.*;
+import static org.junit.Assert.*;
+import static org.hamcrest.CoreMatchers.*;
+import org.openqa.selenium.*;
+import org.openqa.selenium.htmlunit.HtmlUnitDriver;
+import com.gargoylesoftware.htmlunit.BrowserVersion;
+import org.openqa.selenium.support.ui.Select;
+
+public class SeleniumTest {
+	private HtmlUnitDriver driver;
+	private String baseUrl;
+	private boolean acceptNextAlert = true;
+	private StringBuffer verificationErrors = new StringBuffer();
+
+	@Before
+	public void setUp() throws Exception {
+		driver = new HtmlUnitDriver();
+		baseUrl = "goomba-developmentbranch.herokuapp.com/";
+		driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
+	}
+
+	@Test
+	public void testSelenium() throws Exception {
+		driver.get(baseUrl + "");
+		assertEquals("This should fail", driver.getTitle());
+	}
+
+	@After
+	public void tearDown() throws Exception {
+		driver.quit();
+		String verificationErrorString = verificationErrors.toString();
+		if (!"".equals(verificationErrorString)) {
+			fail(verificationErrorString);
+		}
+	}
+}
